pyproject.toml
[tool.poetry]
name = "faqbot-cc"
version = "0.1.0"
description = "A Discord bot for answering frequently asked questions regarding CC."
authors = ["Wendelstein7 <23213464+Wendelstein7@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/SquidDev-CC/FAQBot-CC"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
discord-py-slash-command = "^1.1.0"
pyyaml = "^5.4.1"
prometheus-client = "^0.10.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
mypy = "^0.800"
autopep8 = "^1.5.4"
"discord.py-stubs" = "^1.6.0"

[build-system]
requires = ["poetry>=0.12"]
uild-backend = "poetry.masonry.api"
